refactor: create a nutrient folder for integrating nutrient related files.
