+++ src/container/App.js
@@ -2,10 +2,10 @@ import { useState } from 'react'
 
 import './App.css'
 import CaloriesForm from '../components/CaloriesForm'
-import NutrientForm from '../components/NutrientForm'
-import MetabolizableEnergyTable from '../components/MetabolizableEnergyTable'
-import DryMatterBasisTable from '../components/DryMatterBasisTable'
-import CalciumToPhosphorusRatioResult from '../components/CalciumToPhosphorusRatioResult'
+import NutrientForm from '../components/nutrient/NutrientForm'
+import MetabolizableEnergyTable from '../components/nutrient/MetabolizableEnergyTable'
+import DryMatterBasisTable from '../components/nutrient/DryMatterBasisTable'
+import CalciumToPhosphorusRatioResult from '../components/nutrient/CalciumToPhosphorusRatioResult'
 import Divider from '@mui/material/Divider'
 import Container from '@mui/material/Container'
